+++ src/main/java/com/ibm/cloud/cloudant/kafka/SinkConnector.java
@@ -11,14 +11,17 @@
  * either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.ibm.cloud.cloudant.kafka.connect;
+package com.ibm.cloud.cloudant.kafka;
 
 import com.ibm.cloud.cloudant.kafka.common.InterfaceConst;
 import com.ibm.cloud.cloudant.kafka.common.utils.JavaCloudantUtil;
+import com.ibm.cloud.cloudant.kafka.connect.CachedClientManager;
+import com.ibm.cloud.cloudant.kafka.connect.SinkConnectorConfig;
+import com.ibm.cloud.cloudant.kafka.connect.ConfigValidator;
+import com.ibm.cloud.cloudant.kafka.connect.SinkTask;
 import org.apache.kafka.common.config.Config;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.Task;
-import org.apache.kafka.connect.source.SourceConnector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,15 +30,21 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class CloudantSourceConnector extends SourceConnector {
+public class SinkConnector extends org.apache.kafka.connect.sink.SinkConnector {
 
-    private static Logger LOG = LoggerFactory.getLogger(CloudantSourceConnector.class);
+    private static Logger LOG = LoggerFactory.getLogger(SinkConnector.class);
 
     private Map<String, String> configProperties;
 
     @Override
     public ConfigDef config() {
-        return CloudantSourceConnectorConfig.CONFIG_DEF;
+        return SinkConnectorConfig.CONFIG_DEF;
+    }
+
+
+    @Override
+    public String version() {
+        return JavaCloudantUtil.VERSION;
     }
 
     @Override
@@ -44,35 +53,32 @@ public class CloudantSourceConnector extends SourceConnector {
     }
 
     @Override
+    public Class<? extends Task> taskClass() {
+        return SinkTask.class;
+    }
+
+    @Override
+    public List<Map<String, String>> taskConfigs(int maxTasks) {
+        List<Map<String, String>> taskConfigs = new ArrayList<>(maxTasks);
+
+        for (int i = 0; i < maxTasks; i++) {
+            Map<String, String> taskProps = new HashMap<>(configProperties);
+            // add task specific properties here (if any)
+            taskProps.put(InterfaceConst.TASK_NUMBER, String.valueOf(i));
+
+            taskConfigs.add(taskProps);
+        }
+        return taskConfigs;
+    }
+
+    @Override
     public void stop() {
         CachedClientManager.removeInstance(configProperties);
     }
 
     @Override
-    public Class<? extends Task> taskClass() {
-        return CloudantSourceTask.class;
-    }
-
-    @Override
-    public List<Map<String, String>> taskConfigs(int maxTasks) {
-        if (maxTasks > 1) {
-            LOG.warn(String.format("tasks.max requested was %d, but only 1 task supported", maxTasks));
-        }
-        List<Map<String, String>> taskConfigs = new ArrayList<>(1);
-        Map<String, String> taskProps = new HashMap<>(configProperties);
-        taskProps.put(InterfaceConst.TASK_NUMBER, String.valueOf(0));
-        taskConfigs.add(taskProps);
-        return taskConfigs;
-    }
-
-    @Override
-    public String version() {
-        return JavaCloudantUtil.VERSION;
-    }
-
-    @Override
     public Config validate(Map<String, String> connectorConfigs) {
-        CloudantConfigValidator validator = new CloudantConfigValidator(connectorConfigs, config());
+        ConfigValidator validator = new ConfigValidator(connectorConfigs, config());
         return validator.validate();
     }