+++ src/main/java/com/ibm/cloud/cloudant/kafka/SourceChangesConnector.java
@@ -11,14 +11,18 @@
  * either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.ibm.cloud.cloudant.kafka.connect;
+package com.ibm.cloud.cloudant.kafka;
 
 import com.ibm.cloud.cloudant.kafka.common.InterfaceConst;
 import com.ibm.cloud.cloudant.kafka.common.utils.JavaCloudantUtil;
+import com.ibm.cloud.cloudant.kafka.connect.CachedClientManager;
+import com.ibm.cloud.cloudant.kafka.connect.SourceChangesConnectorConfig;
+import com.ibm.cloud.cloudant.kafka.connect.ConfigValidator;
+import com.ibm.cloud.cloudant.kafka.connect.SourceChangesTask;
 import org.apache.kafka.common.config.Config;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.Task;
-import org.apache.kafka.connect.sink.SinkConnector;
+import org.apache.kafka.connect.source.SourceConnector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,21 +31,15 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class CloudantSinkConnector extends SinkConnector {
+public class SourceChangesConnector extends SourceConnector {
 
-    private static Logger LOG = LoggerFactory.getLogger(CloudantSinkConnector.class);
+    private static Logger LOG = LoggerFactory.getLogger(SourceChangesConnector.class);
 
     private Map<String, String> configProperties;
 
     @Override
     public ConfigDef config() {
-        return CloudantSinkConnectorConfig.CONFIG_DEF;
-    }
-
-
-    @Override
-    public String version() {
-        return JavaCloudantUtil.VERSION;
+        return SourceChangesConnectorConfig.CONFIG_DEF;
     }
 
     @Override
@@ -50,32 +48,35 @@ public class CloudantSinkConnector extends SinkConnector {
     }
 
     @Override
-    public Class<? extends Task> taskClass() {
-        return CloudantSinkTask.class;
-    }
-
-    @Override
-    public List<Map<String, String>> taskConfigs(int maxTasks) {
-        List<Map<String, String>> taskConfigs = new ArrayList<>(maxTasks);
-
-        for (int i = 0; i < maxTasks; i++) {
-            Map<String, String> taskProps = new HashMap<>(configProperties);
-            // add task specific properties here (if any)
-            taskProps.put(InterfaceConst.TASK_NUMBER, String.valueOf(i));
-
-            taskConfigs.add(taskProps);
-        }
-        return taskConfigs;
-    }
-
-    @Override
     public void stop() {
         CachedClientManager.removeInstance(configProperties);
     }
 
     @Override
+    public Class<? extends Task> taskClass() {
+        return SourceChangesTask.class;
+    }
+
+    @Override
+    public List<Map<String, String>> taskConfigs(int maxTasks) {
+        if (maxTasks > 1) {
+            LOG.warn(String.format("tasks.max requested was %d, but only 1 task supported", maxTasks));
+        }
+        List<Map<String, String>> taskConfigs = new ArrayList<>(1);
+        Map<String, String> taskProps = new HashMap<>(configProperties);
+        taskProps.put(InterfaceConst.TASK_NUMBER, String.valueOf(0));
+        taskConfigs.add(taskProps);
+        return taskConfigs;
+    }
+
+    @Override
+    public String version() {
+        return JavaCloudantUtil.VERSION;
+    }
+
+    @Override
     public Config validate(Map<String, String> connectorConfigs) {
-        CloudantConfigValidator validator = new CloudantConfigValidator(connectorConfigs, config());
+        ConfigValidator validator = new ConfigValidator(connectorConfigs, config());
         return validator.validate();
     }